requirements.txt
pyrogram==2.0.106
tgcrypto
PyYAML
rich
qrcode
loguru
prettytable
pyreadline3

